Fix split-view divider snapping by using CGFloat/NSInteger signatures
The NSSplitView delegate methods (constrainMinCoordinate,
constrainMaxCoordinate) and the AudioSplitView dividerThickness override
were declared with legacy 32-bit types — float and int — instead of the
CGFloat and NSInteger required by the modern AppKit protocol. On 64-bit
arm64, CGFloat is 8 bytes and NSInteger is 8 bytes, so AppKit invoked
these selectors with mismatched type encodings: the proposed coordinate
was read as garbage and the returned value was interpreted as a
wrong-sized number. The visible symptom was that dragging a divider made
it snap to an unintended position.

Switch all three to CGFloat / NSInteger so the type encodings match what
AppKit dispatches.

+++ Sources/AudioSTWindowController.m
@@ -331,12 +331,12 @@
     return YES;
 }
 
-- (float)splitView:(NSSplitView *)sender constrainMinCoordinate:(float)proposedMin ofSubviewAt:(int)offset
+- (CGFloat)splitView:(NSSplitView *)sender constrainMinCoordinate:(CGFloat)proposedMin ofSubviewAt:(NSInteger)offset
 {
     return [sender isVertical]?100:100;
 }
 
-- (float)splitView:(NSSplitView *)sender constrainMaxCoordinate:(float)proposedMax ofSubviewAt:(int)offset
+- (CGFloat)splitView:(NSSplitView *)sender constrainMaxCoordinate:(CGFloat)proposedMax ofSubviewAt:(NSInteger)offset
 {
     return [sender isVertical]?proposedMax-200:proposedMax-100;
 }
